Finally finished adding the IV
Fixed the mixer code so it adds an IV. Had to change the format
of the payload to send an encoded IV but an encrypted message. The
Identifier is just left in string format since those characters
are safe anyway. Then we encode the whole 3 part payload and
send that as the first toAddress to the Jobcoin API.

Also, fix specs and data helpers